requirements.txt
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
distro==1.9.0
numpy==2.3.1
pandas==2.3.1
pdfminer.six==20250506
pdfplumber==0.11.7
openpyxl==3.1.5
pillow==11.3.0
pycparser==2.22
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tabula-py==2.10.0
tzdata==2025.2
